extract MetaInfoTracker interface
